+++ app/representers/stellar_base/deposit_request_representer.rb
@@ -7,7 +7,13 @@ module StellarBase
     property :max_amount
     property :fee_fixed
     property :fee_percent
-    property :extra_info
+    property :extra_info, exec_context: :decorator
+
+    def extra_info
+      ::JSON.parse(represented.extra_info)
+    rescue
+      {}
+    end
 
   end
 end