+++ app/services/stellar_base/determine_extra_info.rb
@@ -1,0 +1,14 @@
+module StellarBase
+  class DetermineExtraInfo
+
+    def self.call(asset_details)
+      info_from =
+        GetCallbackFrom.(asset_details[:extra_info_from])
+
+      if info_from
+        return info_from.(asset_details)
+      end
+    end
+
+  end
+end